% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedup-data.R
\docType{data}
\name{backgroundGene}
\alias{backgroundGene}
\title{Example vector of yeast genes to use as background set for enrichment.
Script to generate example background gene vector
system.file("genes.R", package = "FEDUP")}
\description{
Example vector of yeast genes to use as background set for enrichment.
Script to generate example background gene vector
system.file("genes.R", package = "FEDUP")
}
