% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneDouble}
\alias{geneDouble}
\title{Example list of a background set and two sets of test genes to use for
enrichment analysis.}
\format{
a named list with three vector elements, one common background
gene vector and two test gene vectors.
}
\usage{
data(geneDouble)
}
\description{
Raw Excel data file (Sup Tab 2) is available from
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7566881/}
}
\details{
Script to prepare data
system.file("script", "genes.R", package = "fedup")
}
\keyword{datasets}
