% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneMulti}
\alias{geneMulti}
\title{Example list of a background set and multiple sets of test genes
to use for enrichment analysis.}
\format{
a named list with thirteen vector elements, one common background
gene vector and twelve test gene vectors.
}
\usage{
data(geneMulti)
}
\description{
Raw Excel data file (Sup Tab 2) is available from
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7566881/}
}
\details{
Script to prepare data
system.file("inst", "script", "genes.R", package = "fedup")
}
\keyword{datasets}
