% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSingle}
\alias{geneSingle}
\title{Example list of a background set and single set of test genes to use for
enrichment analysis.}
\format{
a named list with two vector elements, one common background
gene vector and one test gene vector.
}
\usage{
data(geneSingle)
}
\description{
Raw Excel data file (Sup Tab 2) is available from
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7566881/}
}
\details{
Script to prepare data
system.file("inst", "script", "genes.R", package = "fedup")
}
\keyword{datasets}
