% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedup.R
\name{inputObject}
\alias{inputObject}
\title{# Prepares fedup input.}
\usage{
inputObject(test_genes, background_genes, pathways)
}
\arguments{
\item{test_genes}{(char)}

\item{background_genes}{(char)}

\item{pathways}{(char)}
}
\description{
# Prepares fedup input.
}
