% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathwaysTXT}
\alias{pathwaysTXT}
\title{Example list of yeast SAFE terms obtained from a TXT file.}
\format{
a named list of 30 vectors
}
\usage{
pathwaysTXT
}
\description{
Raw data location
system.file("extdata", "SAFE_terms.txt", package = "FEDUP")
}
\details{
Script to prepare data
system.file("data-raw", "pathwaysTXT.R", package = "FEDUP")
}
\keyword{datasets}
