% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathwaysXLSX}
\alias{pathwaysXLSX}
\title{Example list of SAFE terms obtained from a XLSX file.}
\format{
a named list of 317 vectors
}
\usage{
data(pathwaysXLSX)
}
\description{
Raw data file (S5) is available from
\url{https://boonelab.ccbr.utoronto.ca/supplement/costanzo2016/}
}
\details{
Raw data location
system.file("extdata", "SAFE_terms.xlsx", package = "fedup")

Script to prepare data
system.file("script", "pathwaysXLSX.R", package = "fedup")
}
\keyword{datasets}
