% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDotPlot}
\alias{plotDotPlot}
\title{Visualizes pathway enrichment and depletion using ggplot.}
\usage{
plotDotPlot(
  df,
  x_var,
  y_var,
  x_lab = x_var,
  y_lab = NULL,
  p_title = NULL,
  fill_var = NULL,
  fill_col = NULL,
  fill_lab = fill_var,
  size_var = NULL,
  size_lab = size_var
)
}
\arguments{
\item{df}{(data.frame) table with FEDUP enrichment results to plot}

\item{x_var}{(char) x-axis variable (must be a column value in `df`)}

\item{y_var}{(char) y-axis variable (must be a column value in `df`)}

\item{x_lab}{(char) x-axis label (default `x_var`)}

\item{y_lab}{(char) y-axis label (default NULL)}

\item{p_title}{(char) plot title (default NULL)}

\item{fill_var}{(char) point fill variable (default NULL)}

\item{fill_col}{(char) point fill colours (default NULL)}

\item{fill_lab}{(char) point fill label (default `fill_var`)}

\item{size_var}{(char) point size variable (default NULL)}

\item{size_lab}{(char) point size label (default `size_var`)}
}
\value{
ggplot object with the enrichment dot plot
}
\description{
Visualizes pathway enrichment and depletion using ggplot.
}
\examples{
data(testGene)
data(backgroundGene)
data(pathwaysGMT)
fedup_res <- runFedup(testGene, backgroundGene, pathwaysGMT)
fedup_res$log10fdr <- -log10(fedup_res$fdr) # log10-transform FDR
fedup_plot <- fedup_res[head(order(fdr, real_pathway_frac), 20),]
\dontrun{
plotDotPlot(df = fedup_plot,
            x_var = "log10fdr",
            y_var = "pathways",
            x_lab = "-log10(FDR)",
            fill_var = "enrichment",
            fill_lab = "Enrichment",
            size_var = "real_pathway_frac",
            size_lab = "Gene fraction")
}
}
