% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{readPathways}
\alias{readPathways}
\title{Returns a list of pathways from a file.
Currently supports the following file format: GMT, TXT, XLSX}
\usage{
readPathways(
  pathway_file,
  header = TRUE,
  pathway_col,
  gene_col,
  MIN_GENE = 10L,
  MAX_GENE = 500L
)
}
\arguments{
\item{pathway_file}{(char) path to file with pathway annotations}

\item{header}{(logical) whether pathway_file has a header (default = TRUE)}

\item{pathway_col}{(char) column name with pathway identifiers, one pathway per row
use for non-GMT input (eg "Pathway.ID", default = NULL)}

\item{gene_col}{(char) column name with gene identifiers, one gene per row
use for non-GMT input (eg "Gene.ID", default = NULL)}

\item{MIN_GENE}{(integer) minimum number of genes to be considered in a pathway
(default = 10)}

\item{MAX_GENE}{(integer) maximum number of genes to be considered in a pathway
(default = 500)}
}
\value{
a list of vectors with pathways
}
\description{
Returns a list of pathways from a file.
Currently supports the following file format: GMT, TXT, XLSX
}
\examples{
pathways <- readPathways(
 system.file("extdata", "YeastDatabase_GO_gmt.gmt", package = "FEDUP")
)
pathways <- readPathways(
 system.file("extdata", "SAFE_terms.xlsx", package = "FEDUP"),
 header = TRUE, pathway_col = "Enriched.GO.names", gene_col = "ORF.ID"
)
}
