% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedup.R
\name{runFedup}
\alias{runFedup}
\title{Runs gene enrichment and depletion analysis for a list of pathways.}
\usage{
runFedup(testGene, backgroundGene, pathways)
}
\arguments{
\item{testGene}{(char) vector of genes to use as test set.}

\item{backgroundGene}{(char) vector of genes to use as background set.}

\item{pathways}{(list) list of vectors with pathway annotations.}
}
\value{
table of pathway enrichment and depletion results. Rows represent
tested pathways. Columns represent:
\itemize{
    \item pathway -- name of the pathway, corresponds to
        names(\code{pathways});
    \item size -- size of the pathway;
    \item real_frac -- fraction of \code{testGene} members in pathway;
    \item expected_frac -- fraction of \code{backgroundGene} members in
        pathway;
    \item fold_enrichment -- fold enrichment measure,
        evaluates as \code{real_frac} / \code{expected_frac};
    \item status -- indicator that pathway is enriched or depleted for
        \code{testGene} members;
    \item real_gene -- vector of \code{testGene} gene members annotated
        to \code{pathways};
    \item pvalue -- enrichment p-value calculated via Fisher's exact test;
    \item qvalue -- BH-adjusted p-value
}
}
\description{
Runs gene enrichment and depletion analysis for a list of pathways.
}
\examples{
data(testGene)
data(backgroundGene)
data(pathwaysGMT)
fedup_res <- runFedup(testGene, backgroundGene, pathwaysGMT)
}
