% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testGene}
\alias{testGene}
\title{Example vector of human genes to use as test set for enrichment.}
\format{
a character vector with 190 elements (gene IDs)
}
\usage{
testGene
}
\description{
Script to prepare data
system.file("data-raw", "genes.R", package = "FEDUP")
}
\keyword{datasets}
