% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femap.R
\name{writeFemap}
\alias{writeFemap}
\title{Writes an enrichment dataset file for use in Cytoscape EnrichmentMap.}
\usage{
writeFemap(df, resultsFile)
}
\arguments{
\item{df}{(data.frame) table with FEDUP enrichment results.
(see runFedup() for column descriptions)}

\item{resultsFile}{(char) name of output results file.}
}
\value{
table of gene enrichment and depletion results formatted as a
'Generic results file'. Rows represent tested pathways. Columns represent:
\itemize{
    \item pathway -- pathway ID (must match pathway IDs in the GMT file
        provided to plotFemap());
    \item description -- pathway name or description;
    \item pvalue -- enrichment pvalue;
    \item qvalue -- BH-corrected pvalue;
    \item status -- +1 or -1, to identify enriched or depleted pathways
        (+1 maps to red, -1 maps to blue)
}
}
\description{
Writes an enrichment dataset file for use in Cytoscape EnrichmentMap.
}
\examples{
data(testGene)
data(backgroundGene)
data(pathwaysGMT)
fedupRes <- runFedup(testGene, backgroundGene, pathwaysGMT)
resultsFile <- tempfile("fedupRes", fileext = ".txt")
writeFemap(fedupRes, resultsFile)
}
