% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backgroundGene}
\alias{backgroundGene}
\title{Example vector of human genes to use as background set for enrichment.}
\format{
a character vector with 15965 elements (gene IDs)
}
\usage{
backgroundGene
}
\description{
Script to generate data
system.file("data-raw", "genes.R", package = "FEDUP")
}
\keyword{datasets}
