% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedup-data.R
\docType{data}
\name{pathwaysGMT}
\alias{pathwaysGMT}
\title{Example list of human pathway annotations obtained from a GMT file.
Script to generate example pathway list
system.file("pathwaysGMT.R", package = "FEDUP")}
\description{
Example list of human pathway annotations obtained from a GMT file.
Script to generate example pathway list
system.file("pathwaysGMT.R", package = "FEDUP")
}
