% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathwaysXLSX}
\alias{pathwaysXLSX}
\title{Example list of yeast SAFE terms obtained from a XLSX file.}
\format{
a named list of 30 vectors
}
\usage{
pathwaysXLSX
}
\description{
Raw data location
system.file("extdata", "SAFE_terms.xlsx", package = "fedup")
}
\details{
Script to prepare data
system.file("data-raw", "pathwaysXLSX.R", package = "fedup")
}
\keyword{datasets}
