% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDotPlot}
\alias{plotDotPlot}
\title{Visualizes pathway enrichment and depletion using ggplot.}
\usage{
plotDotPlot(
  df,
  xVar,
  yVar,
  xLab = xVar,
  yLab = NULL,
  pTitle = NULL,
  fillVar = NULL,
  fillCol = NULL,
  fillLab = fillVar,
  sizeVar = NULL,
  sizeLab = sizeVar
)
}
\arguments{
\item{df}{(data.frame) table with FEDUP enrichment results to plot.}

\item{xVar}{(char) x-axis variable (must be a column value in \code{df}).}

\item{yVar}{(char) y-axis variable (must be a column value in \code{df}).}

\item{xLab}{(char) x-axis label (default \code{xVar} value).}

\item{yLab}{(char) y-axis label (default NULL).}

\item{pTitle}{(char) plot title (default NULL).}

\item{fillVar}{(char) point fill variable (default NULL).}

\item{fillCol}{(char) point fill colours (default NULL).}

\item{fillLab}{(char) point fill label (default \code{fillVar} value).}

\item{sizeVar}{(char) point size variable (default NULL).}

\item{sizeLab}{(char) point size label (default \code{sizeVar} value).}
}
\value{
object returned from ggplot with the enrichment dot plot.
}
\description{
Visualizes pathway enrichment and depletion using ggplot.
}
\examples{
data(testGene)
data(backgroundGene)
data(pathwaysGMT)
fedup_res <- runFedup(testGene, backgroundGene, pathwaysGMT)
fedup_plot <- fedup_res[which(fedup_res$qvalue < 0.05), ]
fedup_plot$log10qvalue <- -log10(fedup_plot$qvalue + 1e-10)
fedup_plot$pathway <- gsub("\\\\\%.*", "", fedup_plot$pathway)
plotDotPlot(
    df = fedup_plot,
    xVar = "log10qvalue",
    yVar = "pathway",
    xLab = "-log10(Qvalue)",
    fillVar = "status",
    fillLab = "Enrichment\nstatus",
    sizeVar = "fold_enrichment",
    sizeLab = "Fold enrichment"
)
}
