% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femap.R
\name{plotFemap}
\alias{plotFemap}
\title{Draws a network representation of overlaps among pathway enrichment
results using EnrichmentMap (EM) in Cytoscape.}
\usage{
plotFemap(
    gmtFile,
    resultsFolder,
    pvalue = 1,
    qvalue = 1,
    chartData = "NES_VALUE",
    hideNodeLabels = FALSE,
    netName = "generic",
    netFile = "png"
)
}
\arguments{
\item{gmtFile}{(char) absolute path to GMT file (generated via
\link[fedup]{writePathways})}

\item{resultsFolder}{(char) absolute path to folder with fedup results
(generated via \link[fedup]{writeFemap})}

\item{pvalue}{(numeric) pvalue cutoff (value between 0 and 1; default 1)}

\item{qvalue}{(numeric) qvalue cutoff (value between 0 and 1; default 1)}

\item{chartData}{(char) node chart data (one of NES_VALUE, P_VALUE,
FDR_VALUE, PHENOTYPES, DATA_SET, EXPRESSION_SET, or NONE;
default = NES_VALUE)}

\item{hideNodeLabels}{(logical) if TRUE hides the node label in the EM;
cluster labels generated via AutoAnnotate remain visible}

\item{netName}{(char) name for EM in Cytoscape (default generic)}

\item{netFile}{(char) name of output image (supports png, pdf, svg,
jpeg image formats)}
}
\value{
File name of image to which the network is exported and an open
session of Cytoscape (side effect of plotting EM). NULL if Cytoscape
is not running locally.
}
\description{
Draws a network representation of overlaps among pathway enrichment
results using EnrichmentMap (EM) in Cytoscape.
}
\examples{
data(geneDouble)
data(pathwaysGMT)
fedupRes <- runFedup(geneDouble, pathwaysGMT)
resultsFolder <- tempdir()
writeFemap(fedupRes, resultsFolder)
gmtFile <- tempfile("pathwaysGMT", fileext = ".gmt")
writePathways(pathwaysGMT, gmtFile)
netFile <- tempfile("fedup_EM", fileext = ".png")
plotFemap(
    gmtFile = gmtFile,
    resultsFolder = resultsFolder,
    qvalue = 0.05,
    hideNodeLabels = TRUE,
    netName = "fedup_EM",
    netFile = netFile
)
}
