% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femap.R
\name{plotFemap}
\alias{plotFemap}
\title{Plots an EnrichmentMap (EM) in Cytoscape to visualize enriched and depleted gene sets.}
\usage{
plotFemap(gmt_file, results_file, pvalue = 1, qvalue = 1, net_name = "generic")
}
\arguments{
\item{gmt_file}{(char) path to GMT file (must be absolute path)}

\item{results_file}{(char) path to file with enrichment results (must be absolute path)}

\item{pvalue}{(numeric) pvalue cutoff (value between 0 and 1). Gene set
nodes with a p-value lower than the given value will not be included
   in the network. (default = 1)}

\item{qvalue}{(numeric) FDR qvalue cutoff (value between 0 and 1). Gene set
nodes with a p-value lower than the given value will not be included
in the network. (default = 1)}

\item{net_name}{(char) name for network in Cytoscape}
}
\value{
filename of image to which EM is exported. Also side effect of
 plotting the EM in an open session of Cytoscape.
}
\description{
Plots an EnrichmentMap (EM) in Cytoscape to visualize enriched and depleted gene sets.
}
\examples{
\dontrun{
# not run because this function requires Cytoscape to be installed and open
data(testGene)
data(backgroundGene)
data(pathwaysGMT)
gmt_file <- tempfile("pathwaysGMT", fileext = ".gmt")
writePathways(pathwaysGMT, gmt_file)
fedup_res <- runFedup(testGene, backgroundGene, pathwaysGMT)
results_file <- tempfile("fedup_res", fileext = ".txt")
writeFemap(fedup_res, results_file)
plotFemap(
	gmt_file = gmt_file,
	results_file = results_file,
	qvalue = 0.05,
	net_name = "FEDUP_EM"
)}
}
