% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femap.R
\name{plotFemap}
\alias{plotFemap}
\title{Plots an EnrichmentMap in Cytoscape to visualize enriched gene sets}
\usage{
plotFemap(
  gmt_file,
  EM_file,
  EM_group_file,
  output_folder,
  net_name = "generic",
  image_format = "png"
)
}
\arguments{
\item{gmt_file}{(char) file path to GMT file (this needs to be absolute path name).}

\item{EM_file}{(char) file path to results_EM.txt (generated by writeEmapFile.R).}

\item{EM_group_file}{(char) file path to write pathway groupings as determined
by EnrichmentMap.}

\item{output_folder}{(char) file path to write gene set grouping file. This will be
used for between-pathway SNP association analysis.}

\item{net_name}{(char) name for network in Cytoscape.}

\item{image_format}{(char) one of PNG, PDF, SVG, or JPEG.

# Not run because requires Cytoscape to be installed and open.
# plotEmap(gmt_file=gmt_file, EM_file=EM_file, output_folder=enrich_folder,
    net_name="generic", image_format="png").}
}
\value{
Filename of image to which EnrichmentMap is exported. Also side
		effect of plotting the EnrichmentMap in an open session of Cytoscape.
}
\description{
Plots an EnrichmentMap in Cytoscape to visualize enriched gene sets
}
