% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femap.R
\name{plotFemap}
\alias{plotFemap}
\title{Draws a network representation of overlaps among enriched and depleted
pathways using EnrichmentMap (EM) in Cytoscape.}
\usage{
plotFemap(
  gmtFile,
  resultsFile,
  pvalue = 1,
  qvalue = 1,
  netName = "generic",
  netFile = "png"
)
}
\arguments{
\item{gmtFile}{(char) path to GMT file (must be an absolute path).}

\item{resultsFile}{(char) path to file with fedup results
(must be an absolute path).}

\item{pvalue}{(numeric) pvalue cutoff. Pathways with a higher pvalue
will not be included in the EM (value between 0 and 1; default 1).}

\item{qvalue}{(numeric) qvalue cutoff. Pathways with a higher qvalue
will not be included in the EM (value between 0 and 1; default 1).}

\item{netName}{(char) name for EM in Cytoscape (default generic).}

\item{netFile}{(char) name of output image. Supports png, pdf, svg,
jpeg image formats.}
}
\value{
file name of image to which the network is exported. Also side
 effect of plotting the EM in an open session of Cytoscape.
}
\description{
Draws a network representation of overlaps among enriched and depleted
pathways using EnrichmentMap (EM) in Cytoscape.
}
\examples{
\dontrun{
# not run because Cytoscape needs to be installed and open
data(testGene)
data(backgroundGene)
data(pathwaysGMT)
gmtFile <- tempfile("pathwaysGMT", fileext = ".gmt")
fedupRes <- runFedup(testGene, backgroundGene, pathwaysGMT)
resultsFile <- tempfile("fedupRes", fileext = ".txt")
netFile <- tempfile("fedup_EM", fileext = ".png")
writePathways(pathwaysGMT, gmtFile)
writeFemap(fedupRes, resultsFile)
plotFemap(
    gmtFile = gmtFile,
    resultsFile = resultsFile,
    qvalue = 0.05,
    netName = "fedup_EM",
    netFile = netFile
)
}
}
