% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{readPathways}
\alias{readPathways}
\title{Returns a list of pathways from various file formats.
Currently supports the following file format: GMT, TXT, XLSX.}
\usage{
readPathways(
  pathway_file,
  header = FALSE,
  pathway_col = NULL,
  gene_col = NULL,
  MIN_GENE = 1L,
<<<<<<< HEAD
  MAX_GENE = Inf
=======
  MAX_GENE = Inf,
  GO_class = NULL
>>>>>>> 00991f87a2a7f1e9145a31acc65e1f9ed16d9363
)
}
\arguments{
\item{pathway_file}{(char) path to file with pathway annotations.}

\item{header}{(logical) whether \code{pathway_file} has a header
(default FALSE).}

\item{pathway_col}{(char) column name with pathway identifiers.
For use with non-GMT input files (eg "Pathway.ID"; default NULL).}

\item{gene_col}{(char) column name with gene identifiers.
For use with non-GMT input files (eg "Gene.ID"; default NULL).}

\item{MIN_GENE}{(integer) minimum number of genes to be considered
in a pathway (default = 1).}

\item{MAX_GENE}{(integer) maximum number of genes to be considered
in a pathway (default = Inf).}
<<<<<<< HEAD
=======

\item{GO_class}{(char) if GO terms are present in \code{pathway_file},
which class(es) should be subsetted (BP, MF and/or CC, default NULL).}
>>>>>>> 00991f87a2a7f1e9145a31acc65e1f9ed16d9363
}
\value{
a list of vectors with pathway annotations.
}
\description{
Returns a list of pathways from various file formats.
Currently supports the following file format: GMT, TXT, XLSX.
}
\examples{
pathways <- readPathways(
    system.file("extdata", "Human_Reactome_November_17_2020_symbol.gmt",
    package = "FEDUP"), MIN_GENE = 10, MAX_GENE = 500)
pathways <- readPathways(
    system.file("extdata", "SAFE_terms.xlsx", package = "FEDUP"),
    header = TRUE, pathway_col = "Enriched.GO.names", gene_col = "Gene.ID")
}
