% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedup.R
\name{runFedup}
\alias{runFedup}
\title{Runs gene enrichment and depletion analysis for a list of pathways.}
\usage{
runFedup(test_genes, background_genes, pathways)
}
\arguments{
\item{test_genes}{(char)}

\item{background_genes}{(char)}

\item{pathways}{(char) list of gene sets to test for enrichment}
}
\value{
table of gene enrichment and depletion results. Rows represent tested
pathways. Columns represent:
\itemize{
   \item pathway --
   \item
   \item
}
}
\description{
Runs gene enrichment and depletion analysis for a list of pathways.
}
