% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedup-data.R
\docType{data}
\name{testGene}
\alias{testGene}
\title{Example vector of human genes to use as test set for enrichment.
Script to generate example test gene vector
system.file("genes.R", package = "FEDUP")}
\description{
Example vector of human genes to use as test set for enrichment.
Script to generate example test gene vector
system.file("genes.R", package = "FEDUP")
}
